+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.0-alpha.12
+- chore: update lints package constrain to ^6.0.0
+- fix: track changes of path hosted packages
 ## 0.1.0-alpha.11
 - chore: Remove flutter_test dependency from pubspec.yaml
 ## 0.1.0-alpha.10